dom/tests/mochitest/cowl/test_postmessage_tainted.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display"></p>
  <pre id="test">

  <script class="testbody" type="application/javascript">
  SimpleTest.waitForExplicitFinish();

  function ok_wrapper(result, desc) {
    ok(result, desc);
    SimpleTest.finish();
  }

  </script>
  <iframe src="file_postmessage_tainted.html"></iframe>
</pre>
</body>
</html>
